git-svn-id: https://svnsrv.sdu.dk/svn/RobWork/trunk@2571 494c09c7-a67d-49d4-9938-151fb1965be8

+++ RobWork/src/rw/kinematics/StatelessObject.hpp
@@ -18,7 +18,7 @@ namespace kinematics {
     public:
 
         virtual void addStateData(rw::kinematics::StateStructure::Ptr statestructure){
-            for(int i=0;i<_objectDatas.size();i++){
+            for(size_t i=0;i<_objectDatas.size();i++){
                 statestructure->addData(_objectDatas[i]);
             }
         }
